Refactor and consolidate music service architecture
Reorganized the music service by removing `music_service_v2.py` and integrating relevant logic into `music_service.py`. Introduced a unified queueing system (`SongQueue`) with better decoupling of queue and player logic. Updated related commands in `music_cog.py` for improved multi-guild support.
